* Inherit sources from previous exported dependencies.
When creating a notebook package, sources from other previously-exported
notebooks the current one explicitly depends on are copied inside the
Sources directory.

So, if notebook 1 depends on notebook 0, the sources from 0 will be
copied inside 1. If notebook 2 depends on notebook 1, the sources from 1
(which includes the sources from 0) will be copied.

This makes declarations "visible" by simply importing the previous
package, and not the ones where the symbols were declared. In fact,
declarations from notebook 0 used this way in notebook 2 do not
really come from the implementation in 0, but from the copy in 1.

We probably don't need the dependency lines in the SPM definition any
more.